kernel: Decouple from netaddr+asmap
...by splitting the libbitcoinkernel-used portion of timedata.cpp off to
   kernel/timedata.cpp

Notes for reviewers:
    - I strongly recommend reviewing with the following git-diff flags:
          --patience --color-moved=dimmed-zebra
    - See the diff for src/Makefile.am to see the overall effect.
