+++ src/kernel/timedata.cpp
@@ -1,0 +1,28 @@
+// Copyright (c) 2022 The Bitcoin Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#include <kernel/timedata.h>
+
+#include <util/time.h>
+
+Mutex g_timeoffset_mutex;
+int64_t nTimeOffset GUARDED_BY(g_timeoffset_mutex) = 0;
+
+/**
+ * "Never go to sea with two chronometers; take one or three."
+ * Our three time sources are:
+ *  - System clock
+ *  - Median of other nodes clocks
+ *  - The user (asking the user to fix the system clock if the first two disagree)
+ */
+int64_t GetTimeOffset()
+{
+    LOCK(g_timeoffset_mutex);
+    return nTimeOffset;
+}
+
+int64_t GetAdjustedTime()
+{
+    return GetTime() + GetTimeOffset();
+}
